Image-to-Text-description_generator-main/requirements.txt

openai>=1.40.0
torch
transformers>=4.40.0
sentencepiece
Pillow
fastapi
uvicorn[standard]
python-multipart
streamlit